requirements.txt
flask 
flask_mysqldb 
ntlk 
requests 
tensorflow 
mysql-connector-python 
mysqlclient